Passing a C array to python using the gobject-introspection seem to have couple of issues when using pygobject
Everything works fine when the array is passed as an "in" parameter. But when
it is passed as either "inout" or "out" there are couple of issues with
marshaling parameters in c->python as well as python->c.

1. When the length parameter is passed as inout, the length value is not read
correctly (in this case the length is set to the pointer)

2. In the python->c there are again two issues:
a) A GArray is being returned when expecting C array
b) The contents in the array are encapsulated as GIArgument, instead of the
original data type of the array.

https://bugzilla.gnome.org/show_bug.cgi?id=702508