Add load time option to remove deprecated initializers
Add gi.options dictionary to the main gi module. This gives a new
table for optional behavior in overrides and binding machinery
which may break things in a backwards incompatible way.

Add "remove_deprecated_initializers" option used in Gtk overrides
for blocking the loading of class overrides that only contain deprecated
initializers. When enabled, we get about a 12% performance boost when
importing Gtk. This can be enabled as follows:

    import gi
    gi.options['remove_deprecated_initializers'] = True
    from gi.repository import Gtk

https://bugzilla.gnome.org/show_bug.cgi?id=723736